+++ TTemperature.c
@@ -31,7 +31,7 @@ void TEMP_Init(void)
  */
 SYS_STATUS TEMP_GetState(void)
 {
-    WORD adc = ADC_GetTemp();
+    BYTE adc = ADC_GetTemp();
     temperature = compute_temperature_degrees(adc);
     const BYTE *thresholds = MENU_GetTMPThresholds();
     return compute_temperature_state(temperature, thresholds);
